// alagoas-motos/lib/os-types.ts
export interface OSLinha {
  emissao: string; // ISO
  numero_os: string;
  situacao: string;
  tipo_os: string;
  pessoa: string;
  modelo: string;
  placa: string;
  total_desconto: number;
  total_servico: number;
  total_acrescimo: number;
  total_mercadoria: number;
  valor_total: number;
  empresa: string;
  consultor: string;
}